requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
groq>=1.0.0
ddgs>=9.0.0
praw>=7.7.0
python-dotenv>=1.0.0
urllib3>=2.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
